move tests files to internal integration location
