+++ src/account_transaction.rs
@@ -8,9 +8,9 @@ use diesel::prelude::*;
 use diesel::serialize::{Output, ToSql};
 use diesel::sql_types::Varchar;
 
-use crate::{db, PgPool};
+use crate::db;
 use crate::schema::account_transactions;
-use crate::types::{Id, Result, Time};
+use crate::types::{Id, Time};
 
 #[derive(Queryable, Identifiable, PartialEq, Debug)]
 pub struct AccountTransaction {
@@ -22,11 +22,11 @@ pub struct AccountTransaction {
 }
 
 pub struct Repo {
-	db: PgPool
+	db: db::PgPool
 }
 
 impl Repo {
-	pub fn new(db: PgPool) -> Self {
+	pub fn new(db: db::PgPool) -> Self {
 		Repo { db }
 	}
 	
@@ -47,3 +47,40 @@ pub struct NewAccountTransaction<'a> {
 	pub amount: &'a BigDecimal,
 }
 
+#[cfg(test)]
+mod tests {
+	use crate::testutil::*;
+	
+	use super::*;
+	
+	#[test]
+	fn create_account_transaction() {
+		let mut fixture = Fixture::new();
+		let suite = Suite::setup();
+		
+		let sender_account = fixture.account_factory.checking_account(
+			fixture.user_factory.bob().id
+		);
+		let receiver_account = fixture.account_factory.checking_account(
+			fixture.user_factory.lucy().id
+		);
+		
+		let amount = BigDecimal::from(100);
+		
+		let got = suite.account_transaction_repo.create(NewAccountTransaction {
+			sender_id: &sender_account.id,
+			receiver_id: &receiver_account.id,
+			amount: &amount,
+		}).unwrap();
+		
+		let want = AccountTransaction {
+			id: got.id,
+			sender_id: sender_account.id,
+			receiver_id: receiver_account.id,
+			amount,
+			created_at: got.created_at,
+		};
+		
+		assert_eq!(got, want);
+	}
+}
