+++ src/bank/mod.rs
@@ -1,0 +1,8 @@
+mod service;
+mod error;
+#[cfg(test)]
+mod test;
+
+
+
+
