+++ src/types.rs
@@ -4,11 +4,7 @@ use chrono::{Datelike, DateTime, NaiveDate, Utc};
 use diesel::PgConnection;
 use diesel::r2d2::ConnectionManager;
 
-use crate::error;
-
-pub type Result<T> = std::result::Result<T, error::Error>;
 pub type Id = uuid::Uuid;
-pub type PgPool = r2d2::Pool<ConnectionManager<PgConnection>>;
 pub type Time = DateTime<Utc>;
 
 pub type Date = NaiveDate;
