+++ src/user.rs
@@ -1,7 +1,7 @@
 use diesel::PgConnection;
 use diesel::prelude::*;
 
-use crate::{db, PgPool, Result};
+use crate::db;
 use crate::schema;
 use crate::schema::users;
 
@@ -19,11 +19,11 @@ pub struct User {
 }
 
 pub struct Repo {
-	db: PgPool,
+	db: db::PgPool,
 }
 
 impl Repo {
-	pub fn new(db: PgPool) -> Self {
+	pub fn new(db: db::PgPool) -> Self {
 		Repo { db }
 	}
 	
@@ -69,3 +69,51 @@ pub enum FindKey<'a> {
 }
 
 
+#[cfg(test)]
+mod tests {
+	use std::borrow::Borrow;
+	
+	use crate::testutil::*;
+	
+	use super::*;
+	
+	#[test]
+	fn insert_user() {
+		let fixture = Fixture::new();
+		let suite = Suite::setup();
+		let user = suite.user_repo.create_user(NewUser {
+			email: "example@gmail.com",
+			first_name: "Tom",
+			family_name: "Riddle",
+			phone_number: Some("555-5555"),
+		}).unwrap();
+		
+		let got_user = users::table.find(user.id).first::<User>(&fixture.conn()).unwrap();
+		assert_eq!(got_user, user)
+	}
+	
+	#[test]
+	fn find_user_with_key() {
+		let mut fixture = Fixture::new();
+		let user = fixture.user_factory.bob();
+		
+		let suite = Suite::setup();
+		
+		let email = user.email.borrow();
+		let id = user.id;
+		
+		// test cases using various UserKeys
+		let test_cases = vec![
+			FindKey::Email(email),
+			FindKey::ID(id)
+		];
+		
+		
+		for user_key in test_cases {
+			let got = suite.user_repo.find_user(user_key)
+				.expect("found user");
+			
+			assert_eq!(user, got)
+		}
+	}
+}
