--- tests/lib.rs
@@ -1,13 +1,0 @@
-mod common;
-
-mod repos {
-	mod user;
-	mod account;
-	mod bank_transaction;
-	mod account_transaction;
-	mod loan;
-}
-
-mod bank;
-
-
